requirements.txt
langchain
langchain-ollama
langchain-chroma
langchain_community
python-dotenv